requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=0.15.0=pypi_0
anndata=0.7.8=pypi_0
anykeystore=0.2=pypi_0
astunparse=1.6.3=pypi_0
ca-certificates=2022.4.26=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py37h06a4308_0
charset-normalizer=2.1.0=pypi_0
cloudpickle=2.1.0=pypi_0
cryptacular=1.6.2=pypi_0
cycler=0.11.0=pypi_0
dca=0.3.4=pypi_0
defusedxml=0.7.1=pypi_0
flatbuffers=1.12=pypi_0
fonttools=4.34.2=pypi_0
future=0.18.2=pypi_0
gast=0.3.3=pypi_0
google-auth=2.9.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
greenlet=1.1.2=pypi_0
grpcio=1.32.0=pypi_0
h5py=2.10.0=pypi_0
hupper=1.10.3=pypi_0
hyperopt=0.2.7=pypi_0
idna=3.3=pypi_0
igraph=0.9.11=pypi_0
importlib-metadata=4.12.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.1.0=pypi_0
keras=2.4.3=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.3=pypi_0
kopt=0.1.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
leidenalg=0.8.10=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.38.1=pypi_0
markdown=3.3.7=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.2=pypi_0
natsort=8.1.0=pypi_0
ncurses=6.3=h5eee18b_3
networkx=2.6.3=pypi_0
numba=0.55.2=pypi_0
numexpr=2.8.3=pypi_0
numpy=1.19.5=pypi_0
oauthlib=3.2.0=pypi_0
openssl=1.1.1p=h5eee18b_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pandas=1.3.5=pypi_0
pastedeploy=2.1.1=pypi_0
patsy=0.5.2=pypi_0
pbkdf2=1.3=pypi_0
pillow=9.2.0=pypi_0
pip=21.2.2=py37h06a4308_0
plaster=1.0=pypi_0
plaster-pastedeploy=0.7=pypi_0
protobuf=3.19.4=pypi_0
py4j=0.10.9.5=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pynndescent=0.5.7=pypi_0
pyparsing=3.0.9=pypi_0
pyramid=2.0=pypi_0
pyramid-mailer=0.15.1=pypi_0
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
python3-openid=3.2.0=pypi_0
pytz=2022.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
repoze-sendmail=4.4.1=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scanpy=1.8.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.2=pypi_0
session-info=1.0.0=pypi_0
setuptools=61.2.0=py37h06a4308_0
sinfo=0.3.4=pypi_0
six=1.15.0=pypi_0
sqlalchemy=1.4.39=pypi_0
sqlite=3.38.5=hc218d9a_0
statsmodels=0.13.2=pypi_0
stdlib-list=0.8.0=pypi_0
tables=3.7.0=pypi_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.4.4=pypi_0
tensorflow-estimator=2.4.0=pypi_0
termcolor=1.1.0=pypi_0
texttable=1.6.4=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.12.0=pypi_0
torch-geometric=2.0.4=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.14=pypi_0
tqdm=4.64.0=pypi_0
transaction=3.0.1=pypi_0
translationstring=1.4=pypi_0
typing-extensions=3.7.4.3=pypi_0
umap-learn=0.5.3=pypi_0
urllib3=1.26.9=pypi_0
velruse=1.1.1=pypi_0
venusian=3.0.0=pypi_0
webob=1.8.7=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
wtforms=3.0.1=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xlrd=1.2.0=pypi_0
xz=5.2.5=h7f8727e_1
zipp=3.8.0=pypi_0
zlib=1.2.12=h7f8727e_2
zope-deprecation=4.4.0=pypi_0
zope-interface=5.4.0=pypi_0
zope-sqlalchemy=1.6=pypi_0
